moved side bar content to prototype component
