+++ src/pages/ModalPage.js
@@ -6,51 +6,9 @@ import {
   BreadcrumbLinkItem,
   IconHeading,
   IconButton,
-  Section,
-  SummaryList
+  Section
 } from "@baltimorecounty/dotgov-components";
-
-const sideBarContent = [
-  {
-    id: 1,
-    icon: "fas fa-map-marker-alt",
-    heading: "Main Office",
-    content:
-      "<p> 111 West Chesapeake Avenue<br /> Suite 307<br /> Towson, Maryland 21204</p>"
-  },
-
-  {
-    id: 2,
-    icon: "far fa-clock",
-    heading: "Hours:",
-    content: "<p>Monday through Friday<br />8 a.m. to 4:30 p.m.</p>"
-  },
-  {
-    id: 3,
-    icon: "far fa-envelope",
-    heading: "Email:",
-    content: "<p><a href=`#`>publicworks@​baltimorecountymd.gov</a></p>"
-  },
-  {
-    id: 4,
-    icon: "fas fa-mobile-alt",
-    heading: "Phone:",
-    content:
-      "<p><a href=`tel:410-887-6789`>410-887-6789</a></p><p><h3>Emergency Dispatch:</h3></p><p><a href=`tel:410-887-6789`>410-887-6789</a></p>"
-  },
-  {
-    id: 5,
-    icon: "fas fa-exclamation-circle",
-    heading: "BALTCOGO311",
-    content: " <p><a href=`#`>Report a problem</a> online.</p>"
-  },
-  {
-    id: 6,
-    icon: "fas fa-user",
-    heading: "Director:",
-    content: "<p>Steve Walsh, P.E.</p>"
-  }
-];
+import SideBarContent from "../components/SampleSideBarContent";
 
 const DemoModal = props => {
   const {
@@ -120,8 +78,6 @@ const mainContent = () => (
   </React.Fragment>
 );
 
-const sideBar = () => <SummaryList items={sideBarContent} />;
-
 const breadCrumbs = () => (
   <Breadcrumbs>
     <BreadcrumbLinkItem text="Departments" link="/departments" />
@@ -136,7 +92,7 @@ const ModalPage = props => {
       pageTitle="Modals"
       Breadcrumbs={breadCrumbs}
       MainContent={mainContent}
-      SideBar={sideBar}
+      SideBar={SideBarContent}
     />
   );
 };
